refactor(web): ChatTab -> ConvItem
